# docs/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilitiesResponse": {
        "additionalProperties": false,
        "properties": {
          "engines": {
            "items": {
              "$ref": "#/components/schemas/EngineCapability"
            },
            "title": "Engines",
            "type": "array"
          },
          "interfaces": {
            "default": [
              "http",
              "mcp"
            ],
            "items": {
              "enum": [
                "http",
                "mcp"
              ],
              "type": "string"
            },
            "title": "Interfaces",
            "type": "array"
          },
          "limits": {
            "$ref": "#/components/schemas/CapabilityLimits"
          },
          "object_types": {
            "items": {
              "$ref": "#/components/schemas/ObjectType"
            },
            "title": "Object Types",
            "type": "array"
          }
        },
        "required": [
          "engines"
        ],
        "title": "CapabilitiesResponse",
        "type": "object"
      },
...
      "CatalogJobPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogJobDescriptor"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "CatalogJobPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "awaiting_selection",
          "awaiting_owner",
          "ready",
          "completed",
          "completed_with_warnings",
          "partial",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "ObjectType": {
        "enum": [
          "table",
          "procedure",
          "function"
        ],
        "title": "ObjectType",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "SQL Context Pack",
    "version": "2.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/capabilities": {
      "get": {
        "operationId": "capabilities_api_v1_capabilities_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capabilities",
        "x-sqlctx-examples": {
          "parameters": {
            "authorization": "authorization_example"
          },
          "path": "/api/v1/capabilities",
          "request": null,
          "response": {
            "body": {
              "engines": [
                {
                  "engine": "sqlserver",
                  "sqlfluff_dialect": "sqlfluff_dialect_example"
                }
              ]
            },
            "status": "200"
          }
        }
      }
    },
    "/api/v1/catalogs": {
      "get": {
        "operationId": "list_catalogs_api_v1_catalogs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 250,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogJobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Catalogs",
        "x-sqlctx-examples": {
          "parameters": {
            "authorization": "authorization_example",
            "cursor": "cursor_example",
            "limit": 100,
            "status": "queued"
          },
          "path": "/api/v1/catalogs",
          "request": null,
          "response": {
            "body": {
              "items": [
                {
                  "catalog_id": "cat_example",
                  "profile_name": "demo",
                  "request_fingerprint": "sha256:0000000000000000000000000000000000000000000000000000000000000000",
                  "safe_request_summary": {},
                  "status": "queued"
                }
              ],
              "page": {
                "limit": 1,
                "returned": 1
              }
            },
            "status": "200"
          }
        }
      },
      "post": {
        "operationId": "create_catalog_api_v1_catalogs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Catalog",
        "x-sqlctx-examples": {
          "parameters": {
            "Idempotency-Key": "idem_example",
            "authorization": "authorization_example"
          },
          "path": "/api/v1/catalogs",
          "request": {
            "profile": "demo",
            "schemas": [
              "schemas_example"
            ]
          },
          "response": {
            "body": {
              "catalog_id": "cat_example",
              "request_fingerprint": "sha256:0000000000000000000000000000000000000000000000000000000000000000",
              "status": "queued"
            },
            "status": "202"
          }
        }
      }
    },
    "/api/v1/catalogs/{catalog_id}": {
      "delete": {
        "operationId": "delete_catalog_api_v1_catalogs__catalog_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "catalog_id",
            "required": true,
            "schema": {
              "title": "Catalog Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Catalog",
        "x-sqlctx-examples": {
          "parameters": {
            "authorization": "authorization_example",
            "catalog_id": "cat_example"
          },
          "path": "/api/v1/catalogs/{catalog_id}",
          "request": null,
          "response": {
            "body": {
              "deleted": true,
              "target_id": "target_id_example"
            },
            "status": "200"
          }
        }
      },
      "get": {
        "operationId": "catalog_status_api_v1_catalogs__catalog_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "catalog_id",
            "required": true,
            "schema": {
              "title": "Catalog Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Catalog Status",
        "x-sqlctx-examples": {
          "parameters": {
            "authorization": "authorization_example",
            "catalog_id": "cat_example"
          },
          "path": "/api/v1/catalogs/{catalog_id}",
          "request": null,
          "response": {
            "body": {
              "catalog_id": "cat_example",
              "request_fingerprint": "sha256:0000000000000000000000000000000000000000000000000000000000000000",
              "status": "queued"
            },
            "status": "200"
          }
        }
      }
    },
    "/api/v1/catalogs/{catalog_id}/cancel": {
      "post": {
        "operationId": "cancel_catalog_api_v1_catalogs__catalog_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "catalog_id",
            "required": true,
            "schema": {
              "title": "Catalog Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Catalog",
        "x-sqlctx-examples": {
          "parameters": {
            "authorization": "authorization_example",
            "catalog_id": "cat_example"
          },
          "path": "/api/v1/catalogs/{catalog_id}/cancel",
          "request": null,
          "response": {
            "body": {
              "catalog_id": "cat_example",
              "request_fingerprint": "sha256:0000000000000000000000000000000000000000000000000000000000000000",
              "status": "queued"
            },
            "status": "200"
          }
        }
      }
    },
    "/api/v1/catalogs/{catalog_id}/category-preview": {
      "get": {
        "operationId": "category_preview_api_v1_catalogs__catalog_id__category_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "catalog_id",
            "required": true,
            "schema": {
              "title": "Catalog Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 250,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Category Preview",
        "x-sqlctx-examples": {
          "parameters": {
            "authorization": "authorization_example",
            "catalog_id": "cat_example",
            "cursor": "cursor_example",
            "limit": 100
          },
          "path": "/api/v1/catalogs/{catalog_id}/category-preview",
          "request": null,
          "response": {
            "body": {
              "items": [
                {
                  "category": "category_example",
                  "object_count": 1
                }
              ],
              "page": {
                "limit": 1,
                "returned": 1
              }
            },
            "status": "200"
          }
        }
      }
    },
...
  }
}